"Fix" the problem with unary operator and substraction reported in the previous commit.
Apparently, it was not a problem of the minus itself, but the
substraction, it fails operating with some decimals.
In the previous test, I was trying 5.1 - 1.5, and the result which
was 3.6 failed to be compared against 3.6.

I have reproduced the problem in the lua repl:

> a = 5.1 - 1.5
> b = 3
> assert ( b == 3 )
> assert ( a == 3.6 )
stdin:1: assertion failed!
stack traceback:
   [C]: in function 'assert'
   stdin:1: in main chunk
   [C]: ?
>

I might create a ticket on lua for this, but for the time being, I have
changed the test to use different numbers.