Fix complex_test: attempt to call global 'require' ...
Includes in lua seems to be not so straight forward, and just moving
around a couple of lines can make a huge difference.

A couple of commits ago, I made a commit that had wrong order, the
require were after the module statement. That made the following errro
to appear: attempt to call global 'require'
To fix, the error, I just had to put back the requires before the
"module" statment. This fix is included in this commit.