Minor changes to the stub
   + changed default display size from 800x600 to 320x240, to debug
     easily.
   + do some helful printous.

Thanks to this stub, I have manage to execute the formula without having
to wait to fractaljoy. But also, I have had an idea, and is that instead
of printing "stupid" lines with information, i could generate a gnuplot
map instead. In this way, I can debug and see the graphic generated by
the formula in gnuplot in order to debug it.